mig to main file and del tester
